refactor: deduplicate cell comparison logic and fix sorting bugs
Extract shared XLCellComparer from duplicate code in XLRangeColumn and
XLRangeRow. Also extract BuildCopyResult and ThrowIfTableColumn helpers
in XLRangeColumn to eliminate intra-file duplication.

Fixes:
- matchCase flag was inverted in CompareMixedTypes (used OrdinalIgnoreCase
  when matchCase was true)
- XLRangeColumn compared a cell to itself instead of otherCell
- XLRangeColumn missed TimeSpan/DateTime/Number type handling
